@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.12",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/google-maps": "^13.3.9",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/pwa": "^12.1.0",
-    "@angular/router": "^13.4.0",
-    "@angular/service-worker": "^13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/google-maps": "^14.2.7",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/pwa": "^14.2.13",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@capacitor/android": "4.0.1",
     "@capacitor/app": "4.0.1",
     "@capacitor/camera": "4.0.1",
@@ -40,8 +40,8 @@
     "@capacitor/share": "4.0.1",
     "@capacitor/splash-screen": "4.0.1",
     "@capacitor/status-bar": "4.0.1",
-    "@core/ionic-core": "^1.1.9",
-    "@core/web-core": "^1.0.47",
+    "@core/ionic-core": "^1.2.11",
+    "@core/web-core": "^1.0.81",
     "@ionic-native/android-permissions": "^5.35.0",
     "@ionic-native/call-number": "^5.34.0",
     "@ionic-native/camera": "^5.34.0",
@@ -78,6 +78,7 @@
     "date-fns-tz": "^1.3.5",
     "ionic": "^5.4.16",
     "jetifier": "^2.0.0",
+    "json-2-csv": "^5.5.0",
     "ng2-charts": "^3.0.6",
     "plugin": "^0.3.3",
     "rxjs": "^7.8.1",
@@ -87,15 +88,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-eslint/builder": "^13.5.0",
-    "@angular-eslint/eslint-plugin": "~12.0.0",
-    "@angular-eslint/eslint-plugin-template": "~12.0.0",
-    "@angular-eslint/template-parser": "~12.0.0",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler": "~13.4.0",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-eslint/builder": "^14.4.0",
+    "@angular-eslint/eslint-plugin": "^14.4.0",
+    "@angular-eslint/eslint-plugin-template": "^14.4.0",
+    "@angular-eslint/template-parser": "^14.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "4.0.1",
     "@ionic/angular-toolkit": "^4.0.0",
     "@ionic/lab": "^3.2.10",
@@ -119,7 +120,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.6.4"
+    "typescript": "^4.8.4"
   },
   "description": "An Ionic project"
 }
