@@ -1,5 +1,5 @@
 import { Component, Input, OnInit, OnDestroy, ViewChild, ElementRef, EventEmitter, Output, HostListener, NgZone } from '@angular/core';
-import { FormGroup, FormBuilder, Validators, FormArray, FormControl } from "@angular/forms";
+import { UntypedFormGroup, UntypedFormBuilder, Validators, UntypedFormArray, UntypedFormControl } from "@angular/forms";
 import { DatePipe } from '@angular/common'; 
 import { Router } from '@angular/router';
 import { App_googleService, NotificationService, AppDataShareService, AppPermissionService } from '@core/ionic-core';
@@ -19,7 +19,7 @@ import { AndroidpermissionsService } from 'src/app/service/androidpermissions.se
 import { GridSelectionDetailModalComponent } from '../../modal/grid-selection-detail-modal/grid-selection-detail-modal.component';
 // import { GoogleMap, MapType } from '@capacitor/google-maps';
 import { GoogleMap, MapInfoWindow, MapMarker } from '@angular/google-maps';
-import { ApiService, DataShareService, CustomvalidationService, CommonFunctionService, LimsCalculationsService, CommonAppDataShareService, PermissionService, EnvService, CoreFunctionService, StorageService, Common, GridCommonFunctionService, FileHandlerService } from '@core/web-core';
+import { ApiService, DataShareService, CustomvalidationService, CommonFunctionService, LimsCalculationsService, CommonAppDataShareService, PermissionService, EnvService, CoreFunctionService, StorageService, Common, GridCommonFunctionService, FileHandlerService, ApiCallService, FormCreationService, CheckIfService } from '@core/web-core';
 import { Capacitor } from '@capacitor/core';
 
 interface User {
@@ -190,11 +190,11 @@ tinymceConfig = {}
   // newCapMap: GoogleMap;
   apiLoaded: Observable<boolean>;
   
-  ionicForm: FormGroup;
+  ionicForm: UntypedFormGroup;
   defaultDate = "1987-06-30";
   isSubmitted = false;
 
-  templateForm!: FormGroup;
+  templateForm!: UntypedFormGroup;
 
   forms:any={};
   form:any ={};
@@ -329,7 +329,7 @@ tinymceConfig = {}
     // }
 
   constructor(
-    public formBuilder: FormBuilder,
+    public formBuilder: UntypedFormBuilder,
     private datePipe: DatePipe,
     private router: Router,
     private commonFunctionService:CommonFunctionService,
@@ -356,7 +356,10 @@ tinymceConfig = {}
     private appDataShareService: AppDataShareService,
     private appPermissionService: AppPermissionService,
     private gridCommonFunctionService: GridCommonFunctionService,
-    private fileHandlerService: FileHandlerService
+    private fileHandlerService: FileHandlerService,
+    private apiCallService: ApiCallService,
+    private formCreationService: FormCreationService,
+    private checkIfService: CheckIfService
     ) {
 
       // this.mapsApiLoaded();
@@ -545,7 +548,7 @@ tinymceConfig = {}
   private getNextFormById(id: string) {
     const params = "form";
     const criteria = ["_id;eq;" + id + ";STATIC"];
-    const payload = this.commonFunctionService.getPaylodWithCriteria(params, '', criteria, {});
+    const payload = this.apiCallService.getPaylodWithCriteria(params, '', criteria, {});
     this.apiService.GetNestedForm(payload);
   }
   updateRunningData(data:any){
@@ -676,11 +679,11 @@ tinymceConfig = {}
         checkCreatControl = this.templateForm.get(element.field_name);
       }
       if (this.staticData[element.ddn_field] && checkCreatControl.controls && checkCreatControl.controls.length == 0) {
-        let checkArray: FormArray;
+        let checkArray: UntypedFormArray;
         if (element.parent) {
-          checkArray = this.templateForm.get(element.parent).get(element.field_name) as FormArray;
+          checkArray = this.templateForm.get(element.parent).get(element.field_name) as UntypedFormArray;
         } else {
-          checkArray = this.templateForm.get(element.field_name) as FormArray;
+          checkArray = this.templateForm.get(element.field_name) as UntypedFormArray;
         }
         this.staticData[element.ddn_field].forEach((data, i) => {
           if (this.updateMode) {
@@ -700,12 +703,12 @@ tinymceConfig = {}
               }
             }
             if (selected) {
-              checkArray.push(new FormControl(true));
+              checkArray.push(new UntypedFormControl(true));
             } else {
-              checkArray.push(new FormControl(false));
+              checkArray.push(new UntypedFormControl(false));
             }
           } else {
-            checkArray.push(new FormControl(false));
+            checkArray.push(new UntypedFormControl(false));
           }
         });
       }
@@ -767,7 +770,7 @@ tinymceConfig = {}
           this.close();
         }else{
           //this.commonFunctionService.getStaticData();
-          const payload = this.commonFunctionService.commanApiPayload([],this.tableFields,this.formFieldButtons,this.getFormValue(false));
+          const payload = this.apiCallService.commanApiPayload([],this.tableFields,this.formFieldButtons,this.getFormValue(false));
           this.apiService.getStatiData(payload);
         }
         // if(this.isStepper){
@@ -879,16 +882,16 @@ tinymceConfig = {}
         if(!element.hideOnMobile){
           if(element.type == 'pdf_view'){
             const object = this.elements[this.selectedRowIndex];
-            staticModalGroup.push(this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,object))
+            staticModalGroup.push(this.apiCallService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,object))
           } 
           if(element.field_name && element.field_name != ''){             
             switch (element.type) {
               case "list_of_checkbox":
-                this.commonFunctionService.createFormControl(forControl, element, [], "list")
+                this.formCreationService.createFormControl(forControl, element, [], "list")
                 this.checkBoxFieldListValue.push(element);
                 break;
               case "checkbox":
-                this.commonFunctionService.createFormControl(forControl, element, false, "checkbox")
+                this.formCreationService.createFormControl(forControl, element, false, "checkbox")
                 break;
               case "date":
                 let currentYear = new Date().getFullYear();
@@ -937,7 +940,7 @@ tinymceConfig = {}
                   element['minDate'] = minDateToday;
                   element['maxDate'] = maxDateFromToday;
                 }
-                this.commonFunctionService.createFormControl(forControl, element, '', "text")
+                this.formCreationService.createFormControl(forControl, element, '', "text")
                 break; 
               case "daterange":
                 const date_range = {};
@@ -948,10 +951,10 @@ tinymceConfig = {}
                 if (element.list_of_dates.length > 0) {
                   list_of_dates.forEach((data) => {
                     
-                    this.commonFunctionService.createFormControl(date_range, data, '', "text")
+                    this.formCreationService.createFormControl(date_range, data, '', "text")
                   });
                 }
-                this.commonFunctionService.createFormControl(forControl, element, date_range, "group")                                    
+                this.formCreationService.createFormControl(forControl, element, date_range, "group")                                    
                 break;            
               case "list_of_fields":
               case "group_of_fields":
@@ -997,11 +1000,11 @@ tinymceConfig = {}
                         }
                         switch (data.type) {
                           case "list_of_checkbox":
-                            this.commonFunctionService.createFormControl(list_of_fields, modifyData, [], "list")
+                            this.formCreationService.createFormControl(list_of_fields, modifyData, [], "list")
                             this.checkBoxFieldListValue.push(modifyData);
                             break;
                           case "checkbox":
-                              this.commonFunctionService.createFormControl(list_of_fields, modifyData, false, "checkbox")
+                              this.formCreationService.createFormControl(list_of_fields, modifyData, false, "checkbox")
                               break;
                           case "date":
                             let currentYear = new Date().getFullYear();
@@ -1041,17 +1044,17 @@ tinymceConfig = {}
                               data['minDate'] = minDateToday;
                               data['maxDate'] = maxDateFromToday;
                             }                        
-                            this.commonFunctionService.createFormControl(list_of_fields, modifyData, '', "text")
+                            this.formCreationService.createFormControl(list_of_fields, modifyData, '', "text")
                             break;                         
                           default:
-                            this.commonFunctionService.createFormControl(list_of_fields, modifyData, '', "text")
+                            this.formCreationService.createFormControl(list_of_fields, modifyData, '', "text")
                             break;
                         } 
                       }                 
                     }
                   }
                 }
-                this.commonFunctionService.createFormControl(forControl, element, list_of_fields, "group"); 
+                this.formCreationService.createFormControl(forControl, element, list_of_fields, "group"); 
                 if(element.type == 'list_of_fields'){
                   this.list_of_fields.push(element);
                 }               
@@ -1080,14 +1083,14 @@ tinymceConfig = {}
                         if (element.onchange_function && element.onchange_function_param && element.onchange_function_param != ""){
                           this.calculationFieldList.push(element);
                         }
-                        this.commonFunctionService.createFormControl(stepper_of_fields, modifyData, '', "text")
+                        this.formCreationService.createFormControl(stepper_of_fields, modifyData, '', "text")
                         if(data.tree_view_object && data.tree_view_object.field_name != ""){
                           let treeModifyData = JSON.parse(JSON.stringify(data.tree_view_object));                
                           treeModifyData.is_mandatory=false;
-                          this.commonFunctionService.createFormControl(stepper_of_fields, treeModifyData , '', "text")
+                          this.formCreationService.createFormControl(stepper_of_fields, treeModifyData , '', "text")
                         }
                       });
-                      this.commonFunctionService.createFormControl(forControl, step, stepper_of_fields, "group")
+                      this.formCreationService.createFormControl(forControl, step, stepper_of_fields, "group")
                     } 
                   }); 
                   this.isStepper = true;
@@ -1095,11 +1098,11 @@ tinymceConfig = {}
                 break;
               case "pdf_view" : 
                 const object = this.elements[this.selectedRowIndex];
-                staticModalGroup.push(this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,object))
+                staticModalGroup.push(this.apiCallService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,object))
               break;
               case "input_with_uploadfile":
                 element.is_disabled = true;
-                this.commonFunctionService.createFormControl(forControl, element, '', "text")
+                this.formCreationService.createFormControl(forControl, element, '', "text")
                 break;
               case "grid_selection":
                   if(element && element.gridColumns && element.gridColumns.length > 0){
@@ -1120,15 +1123,15 @@ tinymceConfig = {}
                     element['fieldIndex'] = index;
                   }
                   element['showButton'] = this.checkGridSelectionButtonCondition(element,'add');
-                  this.commonFunctionService.createFormControl(forControl, element, '', "text");
+                  this.formCreationService.createFormControl(forControl, element, '', "text");
                 break;
               case "gmap":
               case "gmapview": 
                 this.requestLocationPermission();
-                this.commonFunctionService.createFormControl(forControl, element, '', "text");
+                this.formCreationService.createFormControl(forControl, element, '', "text");
                 break;
               default:
-                this.commonFunctionService.createFormControl(forControl, element, '', "text");
+                this.formCreationService.createFormControl(forControl, element, '', "text");
                 break;
             }
             
@@ -1136,7 +1139,7 @@ tinymceConfig = {}
             if(element.tree_view_object && element.tree_view_object.field_name != ""){
               let treeModifyData = JSON.parse(JSON.stringify(element.tree_view_object));                
               treeModifyData.is_mandatory=false;
-              this.commonFunctionService.createFormControl(forControl, treeModifyData , '', "text")
+              this.formCreationService.createFormControl(forControl, treeModifyData , '', "text")
             }
           }
           //show if handling
@@ -1172,7 +1175,7 @@ tinymceConfig = {}
           if(element.field_name && element.field_name != ''){              
             switch (element.type) {
               case "dropdown":
-                this.commonFunctionService.createFormControl(forControl, element, '', "text")
+                this.formCreationService.createFormControl(forControl, element, '', "text")
                 break;
               default:
                 break;
@@ -1239,7 +1242,7 @@ tinymceConfig = {}
           if(element.field_name == 'account'){
             this.templateForm.get('account').setValue(account);
             if (element.onchange_api_params && element.onchange_call_back_field && !element.do_not_auto_trigger_on_edit) {
-              payload = this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, this.getFormValue(true)) 
+              payload = this.apiCallService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, this.getFormValue(true)) 
               if(element.onchange_api_params.indexOf("FORM_GROUP") >= 0 || element.onchange_api_params.indexOf("QTMP") >= 0){
                 payload["data"]=this.getFormValue(true);
               } 
@@ -1249,7 +1252,7 @@ tinymceConfig = {}
           if(element.field_name == 'contact'){
             this.templateForm.get('contact').setValue(contact);
             if (element.onchange_api_params && element.onchange_call_back_field && !element.do_not_auto_trigger_on_edit) { 
-              payload = this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, this.getFormValue(true)) 
+              payload = this.apiCallService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, this.getFormValue(true)) 
               if(element.onchange_api_params.indexOf("FORM_GROUP") >= 0 || element.onchange_api_params.indexOf("QTMP") >= 0){
                 payload["data"]=this.getFormValue(true);
               }                  
@@ -1262,7 +1265,7 @@ tinymceConfig = {}
                 if(data.field_name == 'account'){
                   this.templateForm.get(stepData.field_name).get('account').setValue(account);
                   if (data.onchange_api_params && data.onchange_call_back_field && !data.do_not_auto_trigger_on_edit) { 
-                    payload = this.commonFunctionService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, this.getFormValue(true)) 
+                    payload = this.apiCallService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, this.getFormValue(true)) 
                     if(data.onchange_api_params.indexOf("FORM_GROUP") >= 0 || data.onchange_api_params.indexOf("QTMP") >= 0){
                       payload["data"]=this.getFormValue(true);
                     }  
@@ -1271,7 +1274,7 @@ tinymceConfig = {}
                 }
                 if(data.field_name == 'contact'){
                   if (data.onchange_api_params && data.onchange_call_back_field && !data.do_not_auto_trigger_on_edit) { 
-                    payload = this.commonFunctionService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, this.getFormValue(true)) 
+                    payload = this.apiCallService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, this.getFormValue(true)) 
                     if(data.onchange_api_params.indexOf("FORM_GROUP") >= 0 || data.onchange_api_params.indexOf("QTMP") >= 0){
                       payload["data"]=this.getFormValue(true);
                     }  
@@ -1291,7 +1294,7 @@ tinymceConfig = {}
           let multiCollection = JSON.parse(JSON.stringify(this.multipleFormCollection));
           formValue = this.commonFunctionService.getFormDataInMultiformCollection(multiCollection,object);
         }
-        let staticModalG = this.commonFunctionService.commanApiPayload([],this.tableFields,this.formFieldButtons,formValue);
+        let staticModalG = this.apiCallService.commanApiPayload([],this.tableFields,this.formFieldButtons,formValue);
         if(staticModalG && staticModalG.length > 0){
           staticModalG.forEach(element => {
             staticModalGroup.push(element);
@@ -1321,7 +1324,7 @@ tinymceConfig = {}
           if(this.form.api_params_criteria && this.form.api_params_criteria != null){
             criteria=this.form.api_params_criteria
           }
-          staticModal.push(this.commonFunctionService.getPaylodWithCriteria(this.form.api_params,this.form.call_back_field,criteria,this.getFormValue(false)))
+          staticModal.push(this.apiCallService.getPaylodWithCriteria(this.form.api_params,this.form.call_back_field,criteria,this.getFormValue(false)))
           
         }
         if(staticModal.length > 0){
@@ -1547,7 +1550,7 @@ tinymceConfig = {}
       }
 
       let list = [];
-      list.push(this.commonFunctionService.getPaylodWithCriteria(tableField.api_params,tableField.call_back_field,tableField.api_params_criteria,this.getFormValue(false)));
+      list.push(this.apiCallService.getPaylodWithCriteria(tableField.api_params,tableField.call_back_field,tableField.api_params_criteria,this.getFormValue(false)));
        payload['data'] = list;
       this.apiService.DynamicApiCall(payload);
       this.saveCallSubscribe();
@@ -1901,24 +1904,24 @@ tinymceConfig = {}
             this.tableFields.forEach(element => {            
               if(element.field_name == key){                
                 Object.keys(this.dataListForUpload[key]).forEach(child =>{
-                  selectedRow[key][child] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[key][child]);
+                  selectedRow[key][child] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[key][child]);
                 })
               }
             });          
           }else{
-              selectedRow[key] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[key]);
+              selectedRow[key] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[key]);
           }
         } else {
           if(this.dataListForUpload[key] && this.dataListForUpload[key] != null && !Array.isArray(this.dataListForUpload[key]) && typeof this.dataListForUpload[key] === "object"){
             this.tableFields.forEach(element => {
               if(element.field_name == key){                
                 Object.keys(this.dataListForUpload[key]).forEach(child =>{
-                  modifyFormValue[key][child] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[key][child]);
+                  modifyFormValue[key][child] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[key][child]);
                 })
               }
             });          
           }else{
-            modifyFormValue[key] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[key]);
+            modifyFormValue[key] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[key]);
           }       
           
         }
@@ -1992,7 +1995,7 @@ tinymceConfig = {}
     if(hasPermission){
       let gridSelectionValidation:any = this.checkGridSelectionMendetory(); 
       if(this.isFormValid() && gridSelectionValidation.status){
-        let checkCustmizedValuValidation = this.commonFunctionService.checkCustmizedValuValidation(this.tableFields,formValue);
+        let checkCustmizedValuValidation = this.checkIfService.checkCustmizedValuValidation(this.tableFields,formValue);
         if(checkCustmizedValuValidation.status){
           if (this.dataSaveInProgress) {
             this.showNotify = true;
@@ -2202,7 +2205,7 @@ tinymceConfig = {}
             const checkCLTFN = element.onchange_api_params.indexOf('CLTFN')
             if(checkFormGroup == -1 && checkCLTFN == -1){
 
-              const payload = this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, formValueWithCustomData)
+              const payload = this.apiCallService.getPaylodWithCriteria(element.onchange_api_params, element.onchange_call_back_field, element.onchange_api_params_criteria, formValueWithCustomData)
               if(element.onchange_api_params.indexOf('QTMP') >= 0){
                 if(element && element.formValueAsObjectForQtmp){
                   payload["data"]=formValue;
@@ -2223,7 +2226,7 @@ tinymceConfig = {}
                         const checkFormGroup = data.onchange_call_back_field.indexOf("FORM_GROUP");
                         if(checkFormGroup == -1){
               
-                          const payload = this.commonFunctionService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, formValueWithCustomData)
+                          const payload = this.apiCallService.getPaylodWithCriteria(data.onchange_api_params, data.onchange_call_back_field, data.onchange_api_params_criteria, formValueWithCustomData)
                           if(data.onchange_api_params.indexOf('QTMP') >= 0){
                             if(element && element.formValueAsObjectForQtmp){
                                payload["data"]=formValue;
@@ -2237,7 +2240,7 @@ tinymceConfig = {}
                       if(data.tree_view_object && data.tree_view_object.field_name != ""){
                         let editeTreeModifyData = JSON.parse(JSON.stringify(data.tree_view_object));
                         if (editeTreeModifyData.onchange_api_params && editeTreeModifyData.onchange_call_back_field) {
-                          staticModal.push(this.commonFunctionService.getPaylodWithCriteria(editeTreeModifyData.onchange_api_params, editeTreeModifyData.onchange_call_back_field, editeTreeModifyData.onchange_api_params_criteria, formValueWithCustomData));
+                          staticModal.push(this.apiCallService.getPaylodWithCriteria(editeTreeModifyData.onchange_api_params, editeTreeModifyData.onchange_call_back_field, editeTreeModifyData.onchange_api_params_criteria, formValueWithCustomData));
                         }
                       }
                     });
@@ -2249,12 +2252,12 @@ tinymceConfig = {}
           if(element.tree_view_object && element.tree_view_object.field_name != ""){
             let editeTreeModifyData = JSON.parse(JSON.stringify(element.tree_view_object));
             if (editeTreeModifyData.onchange_api_params && editeTreeModifyData.onchange_call_back_field) {
-              staticModal.push(this.commonFunctionService.getPaylodWithCriteria(editeTreeModifyData.onchange_api_params, editeTreeModifyData.onchange_call_back_field, editeTreeModifyData.onchange_api_params_criteria, formValueWithCustomData));
+              staticModal.push(this.apiCallService.getPaylodWithCriteria(editeTreeModifyData.onchange_api_params, editeTreeModifyData.onchange_call_back_field, editeTreeModifyData.onchange_api_params_criteria, formValueWithCustomData));
             }
           }
         }
         if(element.type && element.type == 'pdf_view'){
-          staticModal.push(this.commonFunctionService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,formValueWithCustomData))
+          staticModal.push(this.apiCallService.getPaylodWithCriteria(element.onchange_api_params,element.onchange_call_back_field,element.onchange_api_params_criteria,formValueWithCustomData))
         }
       });
     }
@@ -2266,7 +2269,7 @@ tinymceConfig = {}
       let multiCollection = JSON.parse(JSON.stringify(this.multipleFormCollection));
       formValue = this.commonFunctionService.getFormDataInMultiformCollection(multiCollection,object);
     }
-    let staticModalG = this.commonFunctionService.commanApiPayload([],this.tableFields,this.formFieldButtons,formValue);
+    let staticModalG = this.apiCallService.commanApiPayload([],this.tableFields,this.formFieldButtons,formValue);
     if(staticModalG && staticModalG.length > 0){
       staticModalG.forEach(element => {
         staticModal.push(element);
@@ -2277,7 +2280,7 @@ tinymceConfig = {}
     //   if(this.tab.api_params_criteria && this.tab.api_params_criteria != null){
     //     criteria=this.tab.api_params_criteria
     //   }
-    //   staticModal.push(this.commonFunctionService.getPaylodWithCriteria(this.tab.api_params,this.tab.call_back_field,criteria,{}))
+    //   staticModal.push(this.apiCallService.getPaylodWithCriteria(this.tab.api_params,this.tab.call_back_field,criteria,{}))
       
     // }
     if(this.form && this.form.api_params && this.form.api_params != null && this.form.api_params != "" && this.form.api_params != undefined){         
@@ -2291,7 +2294,7 @@ tinymceConfig = {}
       if(this.editedRowIndex > -1){
         formDataObject = formValue;
       }
-      staticModal.push(this.commonFunctionService.getPaylodWithCriteria(this.form.api_params,this.form.call_back_field,criteria,formDataObject))      
+      staticModal.push(this.apiCallService.getPaylodWithCriteria(this.form.api_params,this.form.call_back_field,criteria,formDataObject))      
     }
     this.callStaticData(staticModal);
   }
@@ -2432,7 +2435,7 @@ tinymceConfig = {}
                 case "input_with_uploadfile":
                   if(formValue[element.field_name] != null && formValue[element.field_name] != undefined){
                     this.dataListForUpload[element.field_name] = JSON.parse(JSON.stringify(formValue[element.field_name]));
-                    const value = this.commonFunctionService.modifyFileSetValue(formValue[element.field_name]);
+                    const value = this.fileHandlerService.modifyFileSetValue(formValue[element.field_name]);
                     this.templateForm.controls[element.field_name].setValue(value);
                   }
                   break;
@@ -3335,7 +3338,7 @@ tinymceConfig = {}
     const payload = [];
     const params = field.api_params;
     const criteria = field.api_params_criteria;
-    payload.push(this.commonFunctionService.getPaylodWithCriteria(params, '', criteria, objectValue, field.data_template));
+    payload.push(this.apiCallService.getPaylodWithCriteria(params, '', criteria, objectValue, field.data_template));
     this.apiService.GetTypeaheadData(payload);    
   }
   clearTypeaheadData() {
@@ -3452,7 +3455,7 @@ tinymceConfig = {}
               if(primaryCriteriaList && primaryCriteriaList.length > 0){
                 for (let index = 0; index < primaryCriteriaList.length; index++) {
                   const cri = primaryCriteriaList[index];
-                  alreadyExist = this.commonFunctionService.checkIfCondition(cri,element,field.type);
+                  alreadyExist = this.checkIfService.checkIfCondition(cri,element,field.type);
                   if(alreadyExist){
                     const crList = cri.split("#");
                     switch (crList[1]) {
@@ -3845,7 +3848,7 @@ tinymceConfig = {}
               }
               if(this.dataListForUpload[keyName]){
                 Object.keys(this.dataListForUpload[keyName]).forEach(childkey => {                  
-                  updateCustmizedValue[this.listOfFieldsUpdateIndex][childkey] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[keyName][childkey]);
+                  updateCustmizedValue[this.listOfFieldsUpdateIndex][childkey] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[keyName][childkey]);
                 })
               }
               if (this.checkBoxFieldListValue.length > 0 && Object.keys(this.staticData).length > 0) {
@@ -3897,7 +3900,7 @@ tinymceConfig = {}
               }
               if(this.dataListForUpload[keyName]){
                 Object.keys(this.dataListForUpload[keyName]).forEach(childkey => {                 
-                  listOfFieldData[childkey] = this.commonFunctionService.modifyUploadFiles(this.dataListForUpload[keyName][childkey]);
+                  listOfFieldData[childkey] = this.fileHandlerService.modifyUploadFiles(this.dataListForUpload[keyName][childkey]);
                 })
               }
               if (this.checkBoxFieldListValue.length > 0 && Object.keys(this.staticData).length > 0) {
@@ -4296,8 +4299,8 @@ tinymceConfig = {}
   }
 
   private getDataForNextForm(reqParams,reqCriteria) {    
-    const request = this.commonFunctionService.getDataForGrid(1, {}, { 'name': reqParams }, [], {}, '');
-    const crList = this.commonFunctionService.getCriteriaList(reqCriteria, {});
+    const request = this.apiCallService.getDataForGrid(1, {}, { 'name': reqParams }, [], {}, '');
+    const crList = this.apiCallService.getCriteriaList(reqCriteria, {});
     request.data.crList = crList;
     this.apiService.getNextFormData(request);
   }
@@ -4389,8 +4392,8 @@ tinymceConfig = {}
         const params = child.api_params;
         if(params && params != ''){
           const criteria = ["_id;eq;"+fieldValue._id+";STATIC"]
-          const crList = this.commonFunctionService.getCriteriaList(criteria,{});
-          const payload = this.commonFunctionService.getDataForGrid(1,{},{'name':params},[],{},'');
+          const crList = this.apiCallService.getCriteriaList(criteria,{});
+          const payload = this.apiCallService.getDataForGrid(1,{},{'name':params},[],{},'');
           payload.data.crList = crList;
           this.apiService.getGridData(payload);
           this.updateAddNew = true;
@@ -4444,7 +4447,7 @@ tinymceConfig = {}
         this.updateDataOnFormField(calculatedCost);
       }
       else{
-        staticModal.push(this.checkQtmpApi(params,field,this.commonFunctionService.getPaylodWithCriteria(params, callback, criteria, completeObject, data_template))); 
+        staticModal.push(this.checkQtmpApi(params,field,this.apiCallService.getPaylodWithCriteria(params, callback, criteria, completeObject, data_template))); 
         // staticModal.push(this.coreUtilityService.getPaylodWithCriteria(params, callback, criteria, object))      
         // if(params.indexOf("FORM_GROUP") >= 0 || params.indexOf("QTMP") >= 0){
         //   if(field && field.formValueAsObjectForQtmp){
@@ -4744,7 +4747,7 @@ tinymceConfig = {}
                     });
                   }
                   let criteria = crList[0]+"#"+crList[1]+"#"+value;
-                  check = this.commonFunctionService.checkIfCondition(criteria,object);
+                  check = this.checkIfService.checkIfCondition(criteria,object);
                   if(!check){
                     break;
                   } 
@@ -4870,14 +4873,14 @@ tinymceConfig = {}
     if(this.curFileUploadFieldparentfield != ''){
       const custmizedKey = this.commonFunctionService.custmizedKey(this.curFileUploadFieldparentfield); 
       if(this.dataListForUpload[custmizedKey][this.curFileUploadField.field_name] && this.dataListForUpload[custmizedKey][this.curFileUploadField.field_name].length > 0){
-        let fileName = this.commonFunctionService.modifyFileSetValue(this.dataListForUpload[custmizedKey][this.curFileUploadField.field_name]);        
+        let fileName = this.fileHandlerService.modifyFileSetValue(this.dataListForUpload[custmizedKey][this.curFileUploadField.field_name]);        
         this.templateForm.get(this.curFileUploadFieldparentfield.field_name).get(this.curFileUploadField.field_name).setValue(fileName);
       }else{
         this.templateForm.get(this.curFileUploadFieldparentfield.field_name).get(this.curFileUploadField.field_name).setValue('');
       }
     }else{    
       if(this.dataListForUpload[this.curFileUploadField.field_name] && this.dataListForUpload[this.curFileUploadField.field_name].length > 0){
-        let fileName = this.commonFunctionService.modifyFileSetValue(this.dataListForUpload[this.curFileUploadField.field_name]);
+        let fileName = this.fileHandlerService.modifyFileSetValue(this.dataListForUpload[this.curFileUploadField.field_name]);
         this.templateForm.get(this.curFileUploadField.field_name).setValue(fileName);
       }else{
         this.templateForm.get(this.curFileUploadField.field_name).setValue('');
@@ -4940,7 +4943,7 @@ tinymceConfig = {}
     const  formValue = this.getFormValue(true); 
     let tobedesabled;
     if(parent == ''){
-      tobedesabled = this.commonFunctionService.isDisable(chield,this.updateMode,formValue)
+      tobedesabled = this.checkIfService.isDisable(chield,this.updateMode,formValue)
       if(tobedesabled){
         if(!this.templateForm.get(chield.field_name).disabled){
           this.templateForm.get(chield.field_name).disable()
@@ -4951,7 +4954,7 @@ tinymceConfig = {}
         }        
       }
     }else{
-      tobedesabled = this.commonFunctionService.isDisable(chield,this.updateMode,formValue)
+      tobedesabled = this.checkIfService.isDisable(chield,this.updateMode,formValue)
       if(tobedesabled){
         this.templateForm.get(parent).get(chield.field_name).disable()
       }else{
@@ -5146,7 +5149,7 @@ tinymceConfig = {}
         object[key] = this.templateForm.getRawValue()[key];
       })
     }
-    const display = this.commonFunctionService.showIf(field,object);
+    const display = this.checkIfService.showIf(field,object);
     const modifiedField = JSON.parse(JSON.stringify(field));
     modifiedField['display'] = display; 
     field = modifiedField;
@@ -5293,7 +5296,7 @@ tinymceConfig = {}
     if(this.curTreeViewField && this.curTreeViewField.onchange_function_param && this.curTreeViewField.onchange_function_param != ''){
       if(this.curTreeViewField.onchange_function_param.indexOf('QTMP') >= 0){
         const staticModalGroup = []
-        staticModalGroup.push(this.commonFunctionService.getPaylodWithCriteria(this.curTreeViewField.onchange_function_param,'',[],this.getFormValue(true)));
+        staticModalGroup.push(this.apiCallService.getPaylodWithCriteria(this.curTreeViewField.onchange_function_param,'',[],this.getFormValue(true)));
         //this.commonFunctionService.getStaticData(staticModalGroup);
         this.apiService.getStatiData(staticModalGroup);
       }
@@ -5694,7 +5697,7 @@ tinymceConfig = {}
       if(searchValue == undefined || searchValue == ''){
         searchValue = 'this.pageNo';
       }
-      correctIndex = this.gridCommonFunctionService.getCorrectIndex(object,index,field,data,searchValue);
+      correctIndex = this.commonFunctionService.getCorrectIndex(object,index,field,data,searchValue);
     }    
     this.storeFormDetails("",field,index); 
   }
@@ -6120,7 +6123,7 @@ tinymceConfig = {}
     if(field.api_params_criteria && field.api_params_criteria != ''){
       criteria =  field.api_params_criteria;
     }
-    let staticModalGroup = this.commonFunctionService.getPaylodWithCriteria(field.api_params, call_back_field, criteria, this.typeaheadObjectWithtext ? this.typeaheadObjectWithtext : {});
+    let staticModalGroup = this.apiCallService.getPaylodWithCriteria(field.api_params, call_back_field, criteria, this.typeaheadObjectWithtext ? this.typeaheadObjectWithtext : {});
     staticModal.push(staticModalGroup);
     this.apiService.GetTypeaheadData(staticModal);
 
@@ -6146,7 +6149,7 @@ tinymceConfig = {}
   //     return true;
   //   }
   //   if (field.etc_fields && field.etc_fields.disable_if && field.etc_fields.disable_if != '') {
-  //     return this.commonFunctionService.isDisable(field.etc_fields, updateMode, object);
+  //     return this.checkIfService.isDisable(field.etc_fields, updateMode, object);
   //   }   
   //   return false;
   // }
@@ -6158,7 +6161,7 @@ tinymceConfig = {}
         condition = this.tableFields.disableRowIf;
       }
       if(condition != ''){
-        if(this.commonFunctionService.checkDisableRowIf(condition,data)){
+        if(this.checkIfService.checkDisableRowIf(condition,data)){
           check = true;
         }else{
           check = false;
@@ -6171,7 +6174,7 @@ tinymceConfig = {}
     const data = this.commonFunctionService[field.field_name];
     const condition = field.disableRowIf;
     if(condition){
-      return !this.commonFunctionService.checkDisableRowIf(condition,data);
+      return !this.checkIfService.checkDisableRowIf(condition,data);
     }
     return true;    
   }
@@ -6254,7 +6257,7 @@ tinymceConfig = {}
     }
   }
   getValueForGrid(field,object){
-    return this.commonFunctionService.getValueForGrid(field,object);
+    return this.gridCommonFunctionService.getValueForGrid(field,object);
   }
 
   async delete(index:number,field:any,fieldName:string){
@@ -6331,7 +6334,7 @@ tinymceConfig = {}
                         if (!this.dataListForUpload[custmizedKey]) this.dataListForUpload[custmizedKey] = {};
                         if (!this.dataListForUpload[custmizedKey][data.field_name]) this.dataListForUpload[custmizedKey][data.field_name] = [];
                         this.dataListForUpload[custmizedKey][data.field_name] = JSON.parse(JSON.stringify(object[data.field_name]));
-                        const value = this.commonFunctionService.modifyFileSetValue(object[data.field_name]);
+                        const value = this.fileHandlerService.modifyFileSetValue(object[data.field_name]);
                         this.templateForm.get(element.field_name).get(data.field_name).setValue(value);
                       }
                       break;
@@ -6436,7 +6439,7 @@ tinymceConfig = {}
       for (let index = 0; index < fieldValue.length; index++) {
         const value = fieldValue[index];
         formValue[parent] = value;
-        if(this.commonFunctionService.showIf(field,formValue)){
+        if(this.checkIfService.showIf(field,formValue)){
           check = 1;
           break;
         }        
@@ -6637,7 +6640,7 @@ tinymceConfig = {}
     const  formValue = this.getFormValue(true);   
     let tobedesabled;
     if(parent == ''){
-      tobedesabled = this.commonFunctionService.isMendetory(chield,formValue)
+      tobedesabled = this.checkIfService.isMendetory(chield,formValue)
       if(tobedesabled){
         if(this.templateFormControl[chield.field_name].status == 'VALID'){
           this.templateForm.get(chield.field_name).setValidators([Validators.required]);
@@ -6650,7 +6653,7 @@ tinymceConfig = {}
         }        
       }
     }else{
-      tobedesabled = this.commonFunctionService.isMendetory(chield,formValue)
+      tobedesabled = this.checkIfService.isMendetory(chield,formValue)
       if(tobedesabled){
         if(this.templateFormControl[parent][chield.field_name].status == 'VALID'){
           this.templateForm.get(parent).get(chield.field_name).setValidators([Validators.required]);
@@ -6693,7 +6696,7 @@ tinymceConfig = {}
       let api_params = field.onClickApiParams;
       let callBackfield = field.onClickCallBackField;
       let criteria = field.onClickApiParamsCriteria
-      const payload = this.commonFunctionService.getPaylodWithCriteria(api_params,callBackfield,criteria,this.getFormValue(false));
+      const payload = this.apiCallService.getPaylodWithCriteria(api_params,callBackfield,criteria,this.getFormValue(false));
       let payloads = [];
       payloads.push(this.checkQtmpApi(api_params,field,payload));
       this.apiService.getStatiData(payloads);
