@@ -2,7 +2,7 @@ import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { NotificationService } from '@core/ionic-core';
 import { ActionSheetController, ModalController, Platform } from '@ionic/angular';
 import { Camera, CameraResultType, CameraSource, ImageOptions, Photo, GalleryImageOptions, GalleryPhoto, GalleryPhotos} from '@capacitor/camera';
-import { ApiService, CommonFunctionService, DataShareService, LimsCalculationsService, CoreFunctionService, StorageService, GridCommonFunctionService, FileHandlerService } from '@core/web-core';
+import { ApiService, CommonFunctionService, DataShareService, LimsCalculationsService, CoreFunctionService, StorageService, GridCommonFunctionService, FileHandlerService, CheckIfService, ApiCallService } from '@core/web-core';
 import { DatePipe } from '@angular/common';
 import { Filesystem, Directory } from '@capacitor/filesystem';
 import { DataShareServiceService } from 'src/app/service/data-share-service.service';
@@ -68,7 +68,6 @@ export class GridSelectionDetailModalComponent implements OnInit {
     private modalController: ModalController,
     private dataShareService: DataShareService,
     private apiService: ApiService,
-    private datePipe: DatePipe,
     private notificationService: NotificationService,
     private coreFunctionService: CoreFunctionService,
     private commonFunctionService: CommonFunctionService,
@@ -77,7 +76,10 @@ export class GridSelectionDetailModalComponent implements OnInit {
     private storageService: StorageService,
     private plt: Platform,
     private gridCommonFunctionService: GridCommonFunctionService,
-    private dataShareServiceService: DataShareServiceService
+    private dataShareServiceService: DataShareServiceService,
+    private datePipe: DatePipe,
+    private checkIfService: CheckIfService,
+    private apiCallService: ApiCallService
   ) { 
     this.staticDataSubscription = this.dataShareService.staticData.subscribe(data =>{
       this.setStaticData(data);
@@ -149,6 +151,22 @@ export class GridSelectionDetailModalComponent implements OnInit {
     //   this.gridData = [];
     // }
     if (this.field.gridColumns && this.field.gridColumns.length > 0) {
+      let gridColumns = this.commonFunctionService.updateFieldInList('display',this.field.gridColumns);
+      gridColumns.forEach(field => {
+        if (this.coreFunctionService.isNotBlank(field.show_if)) {
+          if (!this.checkIfService.showIf(field, parentObject)) {
+            field['display'] = false;
+          } else {
+            field['display'] = true;
+          }
+        } else {
+          field['display'] = true;
+        }
+        if(field['field_class']){
+          field['field_class'] = field['field_class'].trim();
+        }
+      });
+      this.columnList = gridColumns;
       // let gridColumns = this.commonFunctionService.updateFieldInList('display',this.field.gridColumns);
       // gridColumns.forEach(field => {
       //   if (this.coreFunctionService.isNotBlank(field.show_if)) {
@@ -216,13 +234,13 @@ export class GridSelectionDetailModalComponent implements OnInit {
   checkRowDisabledIf(field,data){
     const condition = field.disableRowIf;
     if(condition){
-      return !this.commonFunctionService.checkDisableRowIf(condition,data);
+      return !this.checkIfService.checkDisableRowIf(condition,data);
     }
     return true;    
   }
   getStaticDataWithDependentData(){
     const staticModal = []
-    let staticModalGroup = this.commonFunctionService.commanApiPayload([],this.columnList,[],this.data);
+    let staticModalGroup = this.apiCallService.commanApiPayload([],this.columnList,[],this.data);
     if(staticModalGroup.length > 0){
       staticModalGroup.forEach(element => {
         staticModal.push(element);
@@ -292,7 +310,7 @@ export class GridSelectionDetailModalComponent implements OnInit {
       return true;
     }
     if (field.etc_fields && field.etc_fields.disable_if && field.etc_fields.disable_if != '') {
-      return this.commonFunctionService.isDisable(field.etc_fields, updateMode, object);
+      return this.checkIfService.isDisable(field.etc_fields, updateMode, object);
     }   
     return false;
   }
@@ -304,7 +322,7 @@ export class GridSelectionDetailModalComponent implements OnInit {
         condition = this.field.disableRowIf;
       }
       if(condition != ''){
-        if(this.commonFunctionService.checkDisableRowIf(condition,data)){
+        if(this.checkIfService.checkDisableRowIf(condition,data)){
           check = true;
         }else{
           check = false;
@@ -345,7 +363,7 @@ export class GridSelectionDetailModalComponent implements OnInit {
     if(field.api_params_criteria && field.api_params_criteria != ''){
       criteria =  field.api_params_criteria;
     }
-    let staticModalGroup = this.commonFunctionService.getPaylodWithCriteria(field.api_params, call_back_field, criteria, this.typeaheadObjectWithtext ? this.typeaheadObjectWithtext : {});
+    let staticModalGroup = this.apiCallService.getPaylodWithCriteria(field.api_params, call_back_field, criteria, this.typeaheadObjectWithtext ? this.typeaheadObjectWithtext : {});
     staticModal.push(staticModalGroup);
     this.ngLoading = true;
     this.apiService.GetTypeaheadData(staticModal);
@@ -373,7 +391,8 @@ export class GridSelectionDetailModalComponent implements OnInit {
       }
     }
     if(inputSelectValue){
-      if(this.commonFunctionService.checkDataAlreadyAddedInListOrNot("_id",option,alreadyAddedList)){
+      let checkStatus = this.checkIfService.checkDataAlreadyAddedInListOrNot("_id",option,alreadyAddedList)
+      if(checkStatus.status){
         if(typeof option == 'string'){
           this.notificationService.presentToastOnBottom( option + ' Already Added');
         }else{
